moved onGetNetworks function to new file
