+++ src/routes/config.h
@@ -48,6 +48,4 @@ void onGetConfig(AsyncWebServerRequest *request);
 
 void onPutConfig(AsyncWebServerRequest *request, uint8_t *data, size_t len, size_t index, size_t total);
 
-void onGetNetworks(AsyncWebServerRequest *request);
-
 #endif
