+++ src/routes/networks.h
@@ -1,0 +1,9 @@
+#include <AsyncWebServer_ESP32_W5500.h>
+#include <ArduinoJson.h>
+
+#ifndef NETWORKS_H
+#define NETWORKS_H
+
+void onGetNetworks(AsyncWebServerRequest *request);
+
+#endif
